1
2

+++ 4.0.1/aspnet-core/src/Infrastructure/UnionMall.EntityFrameworkCore/EntityFrameworkCore/Repositories/SqlExecuter.cs
@@ -48,7 +48,7 @@ namespace UnionMall.EntityFrameworkCore.Repositories
             }
 
         }
-        public DataSet GetCategoryDropDownList(int?tenantId, int parentId = 0, int type = 0)
+        public DataSet GetCategoryDropDownList(int?tenantId, long parentId = 0, int type = 0)
         {
             DataSet ds = new DataSet();
             var configuration = AppConfigurations.Get(WebContentDirectoryFinder.CalculateContentRootFolder());
